Extracted FixtureResponseBuilder from FixturesHandler
